ubah routing dari dashboar menjadi admin

+++ routes/web.php
@@ -18,9 +18,9 @@ use App\Http\Controllers\PaketController;
 //     return view('index');
 // })->name('index');
 
-Route::get('/dashboard', function () {
-    return view('dashboard.index');
-})->name('dashboard');
+Route::get('/admin', function () {
+    return view('admin.index');
+})->name('admin');
 
 Route::get('/register', function () {
     return view('register');
@@ -30,15 +30,28 @@ Route::get('/login', function () {
     return view('login');
 })->name('login');
 
-Route::get('/tambah', function () {
-    return view('dashboard.paket.tambah');
+Route::controller(PaketController::class)->name('paket.')->group(function () {
+    Route::get('/', 'landingpage')->name('landingpage');
+    Route::get('admin/paket', 'index')->name('index');
+    Route::post('admin/paket/simpan', 'create')->name('simpan');
+    Route::get('admin/paket/ubah/{paket}', 'ubah')->name('ubah');
+    Route::patch('admin/paket/update/{paket}', 'update')->name('update');
+    Route::delete('admin/paket/hapus/{paket}', 'delete')->name('delete');
+});
+
+Route::get('/admin/paket/tambah', function () {
+    return view('admin.paket.tambah');
 })->name('paket.tambah');
 
-Route::controller(PaketController::class)->name('paket.')->group(function () {
-    Route::get('dashboard/paket', 'index')->name('index');
-    Route::get('/', 'landingpage')->name('landingpage');
-    Route::post('/simpan', 'create')->name('simpan');
-    Route::get('/ubah/{paket}', 'ubah')->name('ubah');
-    Route::patch('/update/{paket}', 'update')->name('update');
-    Route::delete('/hapus/{paket}', 'delete')->name('delete');
+
+Route::controller(PasananController::class)->name('pesanan.')->group(function () {
+    Route::get('admin/pesanan', 'index')->name('index');
+    Route::post('admin/pesanan/simpan', 'create')->name('simpan');
+    Route::get('admin/pesanan/ubah/{paket}', 'ubah')->name('ubah');
+    Route::patch('admin/pesanan/update/{paket}', 'update')->name('update');
+    Route::delete('admin/pesanan/hapus/{paket}', 'delete')->name('delete');
 });
+
+Route::get('/admin/pesanan/tambah', function () {
+    return view('admin.pesanan.tambah');
+})->name('pesanan.tambah');
